Added org.robolectric:shadows-support-v4 to gradle build. Refactored SwipeRefreshLayout to use support shadow class.
