+++ extras/src/test/java/org/robobinding/supportwidget/swiperefreshlayout/OnRefreshAttributeTest.java
@@ -1,8 +1,10 @@
 package org.robobinding.supportwidget.swiperefreshlayout;
 
 import static org.junit.Assert.assertTrue;
-
-import java.lang.reflect.Field;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -10,6 +12,7 @@ import org.junit.runner.RunWith;
 import org.robobinding.widget.EventCommand;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.RuntimeEnvironment;
+import org.robolectric.Shadows;
 import org.robolectric.annotation.Config;
 
 import android.support.v4.widget.SwipeRefreshLayout;
@@ -22,41 +25,33 @@ import android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener;
  * @author Liang Song
  */
 @RunWith(RobolectricTestRunner.class)
-@Config(manifest = Config.NONE)
+@Config(manifest = Config.NONE, shadows = {ShadowSwipeRefreshLayout.class})
 public class OnRefreshAttributeTest {
 
 	private SwipeRefreshLayout view;
 	private OnRefreshAttribute attribute;
 	private EventCommand eventCommand;
-	private SwipeRefreshLayoutAddOn viewAddOn;
 	
 	@Before
 	public void setUp() throws Exception {
-		//MockitoAnnotations.initMocks(this);
 		view = new SwipeRefreshLayout(RuntimeEnvironment.application);
 		attribute = new OnRefreshAttribute();
 		eventCommand = new EventCommand();
-		viewAddOn = new SwipeRefreshLayoutAddOn(view);
 	}
 	
 	@Test
 	public void givenBoundAttribute_whenPerformOnRefresh_thenEventReceived() throws Exception{
-		bindAttribute();
+		bindAttribute(new SwipeRefreshLayoutAddOn(view));
 		performOnRefresh();
 		assertEventReceived();
 	}
 
-	/*@Test
-	public void whenBinding_thenRegisterWithViewListeners() {
-		bindAttribute();
-		Mockito.when(viewAddOn.addOnRefreshListener(Mockito.any(OnRefreshListener.class))).;
-		Mockito.verify(viewAddOn, Mockito.times(1)).addOnRefreshListener(Mockito.any());
-	}*/
 	private void performOnRefresh() throws Exception{
-		getOnRefreshListener().onRefresh();
+		ShadowSwipeRefreshLayout shadowView = (ShadowSwipeRefreshLayout)Shadows.shadowOf(view);
+		shadowView.getOnRefreshListener().onRefresh();
 	}
 	
-	private void bindAttribute() {
+	private void bindAttribute(SwipeRefreshLayoutAddOn viewAddOn) {
 		attribute.bind(viewAddOn, eventCommand, view);
 	}
 	
@@ -66,10 +61,11 @@ public class OnRefreshAttributeTest {
 		assertTrue(refreshEvent.getView() == view);
 	}
 	
-	private OnRefreshListener getOnRefreshListener() throws Exception {
-		Field f = view.getClass().getDeclaredField("mListener");
-		f.setAccessible(true);
-		OnRefreshListener onRefreshListener = (OnRefreshListener) f.get(view);
-		return onRefreshListener;
+	@Test
+	public void whenBinding_thenRegisterWithViewListeners() {
+		SwipeRefreshLayoutAddOn viewAddOn = mock(SwipeRefreshLayoutAddOn.class);
+		bindAttribute(viewAddOn);
+		
+		verify(viewAddOn, times(1)).addOnRefreshListener(any(OnRefreshListener.class));
 	}
 }
