+++ extras/src/test/java/org/robobinding/supportwidget/swiperefreshlayout/SwipeRefreshAddOnTest.java
@@ -1,7 +1,5 @@
 package org.robobinding.supportwidget.swiperefreshlayout;
 
-import java.lang.reflect.Field;
-
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -10,6 +8,7 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.RuntimeEnvironment;
+import org.robolectric.Shadows;
 import org.robolectric.annotation.Config;
 
 import android.support.v4.widget.SwipeRefreshLayout;
@@ -21,7 +20,7 @@ import android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener;
  * @version
  * @author Liang Song
  */
-@Config(manifest = Config.NONE)
+@Config(manifest = Config.NONE, shadows = {ShadowSwipeRefreshLayout.class})
 @RunWith(RobolectricTestRunner.class)
 public class SwipeRefreshAddOnTest {
 	private SwipeRefreshLayout view;
@@ -44,17 +43,11 @@ public class SwipeRefreshAddOnTest {
 	public void shouldSupportMultipleOnRefreshListeners() throws Exception {
 		viewListeners.addOnRefreshListener(listener1);
 		viewListeners.addOnRefreshListener(listener2);
-		getOnRefreshListener().onRefresh();
+		
+		ShadowSwipeRefreshLayout shadowView = (ShadowSwipeRefreshLayout)Shadows.shadowOf(view);
+		shadowView.getOnRefreshListener().onRefresh();
+		
 		Mockito.verify(listener1, Mockito.times(1)).onRefresh();
 		Mockito.verify(listener2, Mockito.times(1)).onRefresh();
-	}
-
-	// there is no way to trigger OnRefreshListener.onRefresh() in
-	// SwipeRefreshLayout programmatically, use reflection to call.
-	private OnRefreshListener getOnRefreshListener() throws Exception {
-		Field f = view.getClass().getDeclaredField("mListener");
-		f.setAccessible(true);
-		OnRefreshListener onRefreshListener = (OnRefreshListener) f.get(view);
-		return onRefreshListener;
 	}
 }
