+++ framework/src/test/java/org/robobinding/widget/AbstractMultiTypePropertyViewAttributeTest.java
